refactor, working version of trainer_non_syn_vae
