+++ .gitignore
@@ -1,1 +1,5 @@
 .idea
+models/__pycache__/
+outputs/
+trainers/__pycache__/
+utils/__pycache__/
