+++ test/test_metric_factorVAE.py
