+++ trainers/train_non_syn_vae.py
@@ -5,9 +5,9 @@ import torch
 import torch.optim as optim
 import os
 
-from ops import recon_loss, kl_div, kl_div_uni_dim
-from utils import traverse
-from model import VAE
+from utils.ops import recon_loss, kl_div, kl_div_uni_dim
+from utils.utils import traverse
+from models.model_vae import VAE
 from test_plot_gt import plot_gt_shapes
 
 from test import greedy_policy_Smax_discount_worst
